Polidea: Added capabilities and datatypes
git-svn-id: https://redmine.natlab.research.philips.com/svnroot/moonshine/trunk/03_software/Source/Android/Moonshine/shinelib@4512 3c2db180-f856-4c6d-be1c-585a5d938589

+++ src/main/java/com/pins/philips/shinelib/datatypes/SHNDataEnergyExpenditure.java
@@ -4,13 +4,13 @@ package com.pins.philips.shinelib.datatypes;
  * Created by 310188215 on 13/05/15.
  */
 public class SHNDataEnergyExpenditure extends SHNData {
-    private final int energyExpenditure;
+    private final double energyExpenditure;
 
-    public SHNDataEnergyExpenditure(int energyExpenditure) {
+    public SHNDataEnergyExpenditure(double energyExpenditure) {
         this.energyExpenditure = energyExpenditure;
     }
 
-    public int getEnergyExpenditure() {
+    public double getEnergyExpenditure() {
         return energyExpenditure;
     }
